Only choke a WARN log when distribution url's scheme is indeed http
Motivation:

It's possible to pass a local filesystem url for the distribution url and work offline. When passing such a local url, maven-wrapper chokes a WARN log stating that the distribution url is unsecured. This statement is wrong.

Modification:

Only log a WARN message when distribution scheme is `http`, not different from `https`.